[svn r16] Updated test_fido to be a proper stress tester of Hippo.  It keeps the hippo app alive between calls (it is instantiated at import time) and just adds a new canvas for every new path.  This is probably bad, and memory-hogging, but for the next two hours I don't care.  (However, at 2:20PM, you can be sure I will care.  Why 2:20PM?  Who knows!)
I'll be working with Paul later today to make sure that this can work.  Right
now it seems to hang after a couple data dumps as it attempts to save an image;
almost certainly this is a multti-threading lock situation.  I properf
off-screen rendering comes together, this will cease to be a problem.