[svn r78] fixes for namespace.
added recompilation module.  not super-useful yet.  but you just wait...  it
WILL BE!  (the long-term plan here is to add a 'waf' script to the enzo source
distribution, then integrate waf with yt, which will let us just output to the
nearest directory, etc.)

but, for reference, here is a sample.  in my ~/.yt/config i have:

[yt]
enzo_source_tree: /u/ki/mturk/Research/enzo

and then if i run:

frecomp -p . -j8

it will recompile, not cleaning, on 8 procs, and then toss the resultant file
(whatever it is called) into the current directory.  there's more that can be
done, and ultimately with waf it will be done, but for now this is somewhat
useful, but not life-changing.