[svn r100] ugh.
couple fixes.

one, i fixed the namespace for the sqrt function.

another, and this is a big one, i fixed the call to 'sum' in NewProj, so that
it called the numarray sum, instead of whatever 'sum' by itself refers to.  it
seems that 'sum' by itself has different major/minor from the numarray sum, so
the axes got all messed up.  now newproj should work.

sorry about that, ji-hoon.  but, it should work now.

also, some minor additions to the cluster file output reader.