[svn r176] Lots of changes.  Some good, some probably breakage-inducing.
I have fixed up the MPL backend.  It now works properly with slices and
projections (as near as I can tell) and it has support for two-phase and
three-phase diagrams.  In the future more features will be added to these.

EnzoOutput has been added as a type.  This is a build-up to including the movie
format.  .hierarchy is now a property of EnzoStaticOutput (and it will be of
EnzoOutput, but depending on the type it will be differently reconstructed on
the fly.)

I suspect much of fido and whatnot will be broken.  I am primarily committing
to move over to an environment with easier testing facilities (red).

I've begun writing unit testing for the toolkit, so eventually, I will be able
to check things much more automatically.

Data is now serialized in pytables reliably.