[svn r190] Rewriting fido.  All old functions gone; new functions installed.  Should be more modularized, and should also have less baggage from other modules and whatnot.
In the future, numpy inclusion may be removed, as it brings quite a bit of
baggage with it.  (However, it's used in OutputCollection.py for now, so it's
around for the time being.)

The Watcher function is now handled by a class called Watcher.  Examples to
follow, along with appropriate standalone functions.

I've tested this with a simple CollapseTest run, and it seems to keep up pretty
well.  The warning about using shutil will go away at some point.  I apologize,
but fido is probably a bit slower now, because I wanted it to work with both
BSD and GNU -style fileutils, so I couldn't drop to shell and use xargs or mv.

Look for unit tests coming soon, along with examples in the example directory,
and updated scripts in the scripts directory.