[svn r250] Started laying groundwork to make fido useful for UCSD-Enzo.
The major change, though, is the phase plots.  ThreePhasePlot and TwoPhasePlot
no longer exist; there is now just PhasePlot, which can have the z be the field
"CellsPerBin".  This recreates TwoPhasePlot.  Additionally, and coolestly, the
histogramming is now done via weave.inline -- so the first time the code gets
run, it'll toss some errors, compile some code, but then all subsequent times
it runs super-fast.  So that's a plus.