[svn r278] Refactored plot pages to subclass PlotPage.
PhasePlots now somewhat useful - info can be modified, and they can be saved.
Next up will be x and y labels.  Also need to set up data-querying for
statusbar updates.

Note that the mouse behavior changed with this; I changed the ClickProcess call
from the release of a mouse button to the depressing of a mouse button.  Not
sure why, but that fixed bugs I was getting with the notebook overriding my
change-of-tabs.