[svn r384] Okay, this one might break things, but I sure hope it doesn't.
 * I have rewritten the HDF5 IO to be lighter-weight than previously, and less
   dependent on PyTables and the attendent overhead.
 * Projections now use a single entry-point to a PointCombine C function.
   Additionally, they have been slightly rewritten to avoid modifying the grid
   objects, and to store all values locally instead.
 * Projections can now project multiple fields at once.
 * Speed improvements in projections.
 * The field 'ones' no longer requires a read of "Density."  This is a model I
   should apply elsewhere.
 * Tests include a couple modifications, and checking for projection
   correctness when projecting multiple fields.

The things to watch out for, in terms of breakage, are all with the compilation.  I think.  I have tested it quite a bit with both unit tests and real-world tests.

Also, right now, it's only the full-grid reading that relies on my new HDF5 wrapper.  I will be working on making slices work with it as well.  That has the potential to speed up slicing on HDF5 *considerably.*