[svn r476] Minor fix to call _redraw_image in the cutting plane image.
Added -- FINALLY -- access to the callbacks for grid boundaries and velocities.

Vector callbacks in *general* ... I'm not so sure about.  They are available,
but I don't know if I'm going to put them in to the menu item.

Contours, for sure.

I'm finally getting around to exposing some of the functionality that's been in
the code for months and months and months...