[svn r555] Couple fixes to the HOP stuff -- first off, previously we were not correctly correcting for the possible presence of stars.  Now the particles you are returned are definitely *just* the ones that HOP was run on.
I removed the superfluous .next(self) method in the hop list.

Calculation of the radius now accounts for periodicity.

The center of mass is NOT accounted for with periodicity.

I fixed some SHOW STOPPING BUGS in the center_of_mass calculation and the
bulk_velocity calculation.  Oops.

I've also added a HaloMassesPositionPlot object.  (This should probably get a
new name.)  It accepts a hop_results list and plots the halo positions and
radii, true-to-life.  I'll add labels another day.

I've also restricted tags to a 32 bit integer on the backend, which fixed some
issues with bincount.  (If you have more than INT_MAX halos, yowza!)