[svn r654] Added docstrings to all of the callbacks.
Fixed a subtle bug with the way cutting plane plots work.  The .transpose()
call has been appended to the buffer received from CPixelize, which brings it
in line with the Pixelize calls.  Additionally, this has been applied to the
quivering results.  I believe it is now correct.

Additionally, removed the cmap=None arg/defaultarg from the Profile1D plots.