[svn r771] New feature, which will be expanding: idioms and recipes designed for ease of use.
I've added a new simple feature to do frame inspection looking for the variable
'pf'.  So you can now do something like this:

from yt.mods import *
pf = EnzoStaticOutput("RD0035")
hop_plot()

and it will pick up the pf and go to town making up projections with Hop halo
positions as circles.  It will output both the Hop stuff and additionally the
projections with Hop circles overlaid.

This is a bit of magic, and I'm not sure it's going to stick around.  But it is
promising for a bit more ease-of-use.  I was told today that I need better
'test cases' of people who have a hard time using YT, and that I should target
these people.  I think that making the syntax more *verb* like is a good step
toward that.  People don't always want to generate lots of EnzoStaticOutputs,
or PlotCollections; often they just want stuff dumped out.  I'm going to try to
make that easier.