[svn r797] Fixed multiple fields in the Octree
Added EnzoRay, which is an arbitrary ray caster, accepting start&end points.

This code sports a warning, because I've only tested it in the orthogonal case
(where it gives alrighty results) and also because I have recruited somebody
else to help out with that.

Additionally, I found that OrthoRay wasn't returning any points if it was on a
grid boundary, which I have fixed by giving the tie-breaker to the Left side.

I'm still learning Cython, so all these Cython routines end up being a lot
faster.