[svn r978] Projections now store their data differently inside the .yt file.  It works now for storing multiple fields for a single axis, it works for adding additional fields to an existing projection, and it works for adding weight fields.
Plots now append the weight field to the plot name and use correct units if you
are taking a weighted projection.

I think serialization of slices and cutting planes is now broken.  I'll take a
look at this later, but it's low priority since I think few if any people use
that.