[svn r987] Added halo mask generator to cut out circles centered on halos with radius of the virial from light cones.  This can be called as a stand-alone or from within the light cone generator.  An addition to the sample script will follow.
Added various improvements to the main light cone projection routine, including
removing redundant calculations of various boolean masks used in the shifting
procedure.  The depth problem is now handle by performing field_cuts on the
line of site coordinate field and corresponding dx, instead of extracting a
periodic region and throwing that as a source to the raven projection routine.
This will allow the light cones to be run in parallel and has also proven to
be much faster than the original method.