[svn r1071] Fixed bug in light cone generator where line of sight field cuts from all previous slices were being compounded on top of one another.  This led to light cone images with progressively less information in them until the code died from a projection with no elements.  This only affected projections done after the switch was made from using periodic regions to field cuts to do the cutting along the line of sight.  Additionally, because of the way the field_cuts array was being passed around in the light cone generator, this bug only showed up for light cones that had additional user-specified field cuts.
Finally, the ProjectLightCone routine now accepted a keyword, 'node,' where
the user can specify a naming keyword to be used along with other information
for serializing the projection object.  This will allow the light cone
generator to tell the difference between two identical random realizations with
different field cuts.  The argument given with the node keyword should be a
string.