[svn r1118] Couple sets of changes I missed last time:
Profiles now respect the bins better.  I digitization is now done correctly so that none of the bins are empty; this might lead to minor mismatches with previous datasets, but I believe it is more accurate now.

_corners on AMRGridPatch is now a property, since it's not used very much but
is quite heavy.  Something will happen with the hierarchy's gridCorners as
well.

Projections now give up the impression that pdx and pdy are different (they
should be) for the sake of fewer arrays in humongous datasets.  This pains me,
but it's a necessary evil for now.