[svn r1131] Clump methods, pass_down and add_info_item, now check to see if their arugments (only the first one in the case of add_info_item) are callable.  If so, they are called.  If not, they are exec'ed or eval'ed.
Added a recipes dictionary to Clump.py with an entry for adding an info item
that is the distance from a clump to the center of mass of the top level
object.  The entry itself is a fuction that called with the master clump as an
argument and a keyword arg for units.

Example:

recipes['DistanceToMainClump'](master_clump,units='au')

The above line adds the info item to the clump_info dict so that it will appear
when write_clumps or write_clump_hierarchy is called.