[svn r1257] Added a new PlotInterface module that handles getting single plots of slices and projections.  In the yt.mods namespace it is called "plots".  The override argument in save_image is now by default off.
The following commands now work:

from yt.mods import *
plots.get_slice("galaxy1200.dir/galaxy1200", "Density", 0)

and alternatively, the string literal can be replaced with a StaticOutput
object.  Right now it's called 'plots' but maybe it should just be imported
into the local namespace.  Center is guessed if not provided.

Phase spheres and profile sphers and whatnot need to be added.

Once Jae-Joon Lee's multi-axis stuff is included in a released MPL, we'll
support that nicely through this interface.  For now though, this is pretty
effective.

I'm closing #158 now, but phase things need to be included.  I have not yet
because I am not yet sure of the appropriate relationship between these
functions and the PlotCollection.  There is some code dupliation that I might
get rid of.