[svn r1362] Added the keyword, 'serialize' (boolean, default True), to projections allowing one to manually disable serialization or projections.  Serialization still only occurs if _okay_to_serialize is True as well, so setting serialize=True will not serialize projections that weren't being serialized before.
This was added to simplify the solution to the problem of the HaloProfiler
getting stuck at an mpi barrier in parallel mode.  In reality, we do not want
halo profiler projections to be saved anyway, since they are very small and
can be extremely numerous.  Additionally, getting halo profiler projections
to serialize is not so doable as it requires all processors to have write
access to the .yt file which is just asking for trouble.

Finally, committing an unrelated fix to a small bug in writing out hdf5 data
in light cones.  Did this a while ago, forgot to commit it.