[svn r1674] Backport from hg:
 * Optimizations to the pixelization step, which should increase performance by about 40% at most for some pixelizations.
 * Adding an image panner, to be better documented, but which provides us with
the ability to control tile display walls, use Chaco for a "google maps" style
interface to plots, and simplify our overall interaction with variable mesh
plots as a whole.

I'm pretty excited about this image panner, because it's a very cool and fun
way to interact with 2D AMR data.  And it's able to operate remotely, so the
client can run on a local machine (where Chaco may be easier to install, for
intance) getting remote images passed back.  It can also handle "windowing", so
that the pixelization jobs can be distributed across machines -- for instance,
for tile displays!