Adding simple new EnzoFOF halo catalog parser and merger tree calculator. Removing old kd.py module, adding scipy's pure-python kdtree as pykdtree.  This kdtree will do ball-queries between two trees, which is pretty darn useful.
The next step for this catalog parser and merger tree is an overall cleaning,
as well as running the calculation on halo catalogs found with yt.  However,
because EnzoFOF can output inline (with relatively high cadence) we don't
mandate that the catalogs come from StaticOutput outputs just yet.

This will need testing and refinement, as the importing to yt from a standalone
may have had some difficulties.