This commit marks a change in how Cython is used in yt.  Previously we checked in generated sources.  However, this has simply become too burdensome -- the diffs whenever we re-generated the Cython code were simply too large, and this bloated the repository unnecessarily.  This adds support for direct generation and compilation using Cython.
This starts by removing .c and .cpp files generated by Cython and adding the
appropriate information about how to generate the .c and .cpp files to the
setup.py files for RAMSES and the amr_utils directories.  Additionally,
dependencies are included in these setup.py files.

However, to get Cython to play nicely with NumPy, I've had to add in a routine
that I found written by Matthew Brett.  I updated it to include some C++
support.

Additionally, as a safeguard for those systems were Cython is not already
included, I have added a brief routine that checks whether or not Cython is
installed.  If it is not, it attempts to install it.  This will still fail on
systems where superuser privs are required, but that should not be the case in
the majority of yt installations that are not on OSX.  On OSX this will require
sudo-ing the installation.