I have identified the problem with the linked lists for the adaptive ray-tracing, and I believe this fixes it.
The collision detection to identify the next grid into which rays pass is still
an issue, but I believe that the current method (of assuming a radially sorted
progression) is the best option at the moment.  Future versions will improve
upon this; I think an obvious next step is to perform an initial pass inside
integrate_brick, set up a linked list of entries, and then to iterate over only
that linked list.

Anyway, there may still be some major bugs, and it's still somewhat slow
(although it only takes 4 minutes to generate an effective 1650^2 image on my
machine) and I have yet to figure out how to visualize, but at least all the
rays are traveling now.