This change makes the Entropy field accessible from any code, not just Enzo.
We accept an optional field parameter "mu" to compute the mean molecular weight, but if mu is not supplied, simply set it to 1.

Note that even under the original definition, the units are not "erg cm^2" but "erg cm^{3\gamma-3}". This is problematic for a spatially varying gamma (which this does not support) since the units of entropy would vary from place to place. Strictly speaking, the "real" entropy is the logarithm of this quantity plus a constant.