Add image deposition to QuadTree.
This allows QuadTree objects (used in projections) to fill image buffers
directly.  My initial tests show this to be faster in all cases for datasets.
However, this depends strongly on the number of data points in the dataset
versus the number of points in the image buffer.  For large, complex datasets,
this always comes ou in favor of the new object.  As an example, a dataset with
8 levels, 3700 grids and 10 million zones -- small, but not trivial -- was
improved from 19 fps to 32 fps with this new method.

This method is not integrated yet, as the pixelizer is also used in a number of
other places.  However, I hope we can dynamically swap pixelizers to use it.