This replaces pmods.py with a new, different implementation.
The old implementation of pmods.py was based on MPI_Import by Asher Langton.
This is now based on his new implementation, simpler_finder.  The old
implementation broke with the relative imports now used in h5py.  The new
version does not.

There may be a performance regression associated with this.  Additionally, I do
not have the ability to test it on Kraken.  However, it works for me.