I believe this fixes the remaining unit tests.
An issue I am seeing is that this has required a lot of nastiness for the
camera objects and the OffAxisProjection plot.  I believe this will be fixed if
we move to cameras having actual data sources, at which point we no longer
really want the "self.dd" attribute in the DummySource.