Change HEALpix camera to use interpolated projection sampler.
Because the partitioned grids in the HEALpix camera (in contrast to
allsky_projection) use interpolated, vertex-centered values, there was a
size-of-brick issue causing a nasty-looking bug.  This fixes that by
(as-needed) swapping to using the interpolated projection sampler.

It also adds the inner_radius (default 10dx) to avoid multipole moments when
starting in dense regions, and it removes a bit of code that duplicated logic
from the plot_allsky_projection routine.

Fixes #570 .