Adding an "all" mass field for Gadget.
This also includes a fix for getting the most recently used field info, field
detection for tuple-based selection, and a slight improvement to the speed of
IO for particle masses.  A more generic solution is needed for the "all"
particle type, which should probably get pushed into the particle IO handler at
a higher level.