Removing ParticleArrays object.  Convert to using Morton Indices for ParticleOctrees.
This is a major change, and one that is not fully baked; additional changes
will need to be made to the particle frontends themselves, where we need to
convert to generating the Morton index, using a Barnes-Hut type indexing for
the particles themselves, and then doing IO in a different way based on that.

The octree generated for the OWLS dataset is identical to the one generated
previously.