Thread cache arguments through the various frontends.
Note that by default, caching is off, but we turn it on when explicitly asking
for an 'all'.  I think this will replicate the behavior users expect.  However,
one thing to keep a look out for is the relative performance of various
operations, and I think we will also want to consider adding either a heuristic
or a contextmanager that will preserve and traverse simultaneously for the
various coordinates.  This would be helpful, for instance, in simulations like
ARTIO where the traversal could potentially be expensive.

This largely completes the enzodiet, as memory usage is now down considerably
-- although not completely -- in my tests.