Adding a new exception, subclassing YTFieldNotFound
This is for instances when a field could not be generated.  Since this
subclasses YTFieldNotFound, existing machinery for catching that particular
exception will continue to work.  And, we should now have a generic solution to
the problem of fields returning None.  However, this may result in additional
derived fields being swallowed rather than created.