This set of changes finishes periodicity for covering grids.
Note that I have made a handful of modifications to the RegionSelector.  This
is not the only place that changes need to be made; I think we may want to
abstract out some of the logic for checking periodicity and have the base
methods conduct those operations while having selectors themselves conduct
specific operations assuming *no* periodicity.