This enables find_max for RAMSES.
Note that in many cases, find_max already worked!  In fact, if finest_levels
were specified to be false, it would have worked just fine as-is.  But what
this will do is ensure that find_max works in all cases, but specifically those
cases where the maximum "level" of the Octree is not the same as levelmax in
the header file -- as in, those simulations which have not yet refined to their
maximum level yet.  In those cases, no domains would be selected in the
inclusion check for the "all_data" selector, and the max would not be found.

This change makes two distinctions:

ds.h.max_level
ds.max_level

These are no longer the same.  The former is the maximum *reached* level of
refinement, whereas the latter is the maximum *possible* level of refinement.
I believe that this behavior is valid, although I am not terribly fond of it.