Moving the unit subsystem into a new top-level yt.units package.
Right now this is just renaming, but I'd like to add more unit stuff and this
makes more organizational sense, I think.  This also makes it clear that the unit
system is isolated from the rest of yt, except the imports for definitions of
physical quantities which could be easily packaged seperately or scraped from
other sources.