Both units will sometimes be None.  Deal with this.
Due to details of how ndarray subclasses are implemented inside numpy,
specifically the way pretty printing works, we will sometimes get to this point
in the code without ever actually setting a unit for either object, even though
one of them is a YTArray.  This deals with that situation while still preserving
the nice behavior of keeping the correct unit registry around.