Adding SDFIndex from pyslice.
Now a .sindex hangs off the dataset. This is used with an idx0 file to provide
offsets/lengths into the full SDF dataset. This will allow chunked io that is
spatially aware. Currently .sindex is instantiated upon request as a property.
It requires knowledge of the location and level of the idx0 file.

Additionally, I've added the ability to specify the sdf header from an alternate
file in the load command.