We want to iterate in sorted order for data_files.
This fixes a potentially difficult bug, wherein the data_files, and thus the
particles, were being iterated over in a different way.  Recent changes to how
the iteration occurred (and inserting a few additional memory changes, which
can change order-of-sorting for data_files non-deterministically) caused this
to show up as an answer test failure in OWLS.

This adds in a sorting order for particle data files.