Remove some unnecessary error handling code from YTArray.
The ensure_same_dimensions function was originally added to ensure that
we are only doing unit operations between quantities that have the same
dimension.

In the meantime, additional error detection code was added in sanitize_units_add
and sanitize_units_mul for the __add__ and __mul__ magic functions and friends.
For ufuncs, which short-circuity magic functions, we test explicitly for
identical units in binary operations.  This means that while
ensure_same_dimensions *was* being called, it never actually raised an error
because the input had already been validated elsewhere.

This patch removes ensure_same_dimensions since it does no useful work. It also
eliminates a place where we catch a generic RuntimeError. This was parituclarly
pernicious in the case of an infinite recursion error, since python raises that
error as a RuntimeError.