Adding a field plugin for fields defined in my_plugins.py.  Closes #899.
These fields *should* be able to raise errors without causing the whole script
to fail.  It needs to be this way because fields defined in my_plugins may only
make sense for certain datasets, yet the my_plugins.py file shouldn't need to be
edited for each dataset.  The solution is to create a special field plugin that
only contains fields defined in the my_plugins.py file.