Prevent writing from happening in parallel.
This is a rather unsatisfactory solution for a persistent memory leak in the
docs build. The issue is inside sphinx and causes a memory leak during the write
phase of the docs build. Since the issue only happens in parallel, forcing a
erial write phase avoids the issue at the cost of a significantly lengthier
build process.