Be more explicit about constructing bounds for smoothed grids.
This performs a few changes; one of the most important ones is that we now
explicitly compute the bounding box, with a buffer zone, at every level as we
cascadingly update the state.  Additionally, the *only* time we know we will
fill every zone is when we are on level 0, so we only send up the diagnostic
then.

This fixes all the issues I have seen so far.